+++ index.html
@@ -10,7 +10,7 @@
 </head>
 <body>
 
-  <main>
+  <main id="main">
     <div class="promo">
       <h1 class="promo__title visually-hidden">Keksobooking. Кексы по соседству</h1>
       <img src="img/keksobooking.svg" alt="Keksobooking. Кексы по соседству" width="215" height="45">
